+++ elmyra.ip.access.epo/elmyra/ip/access/dpma/__init__.py
@@ -1,0 +1,3 @@
+"""
+Access DPMA patent services
+"""